1rag/requirements_rag.txt
flask==3.0.0
langchain>=0.1.0
langchain-google-genai>=2.0.0
langchain-community>=0.0.20
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
google-generativeai>=0.3.2
chromadb>=0.4.18
pypdf>=3.17.4
numpy>=1.24.0
requests>=2.28.0